+++ android/src/main/kotlin/com/deepanshuchaudhary/pick_or_save/PickOrSavePlugin.kt
@@ -145,6 +145,7 @@ class PickOrSavePlugin : FlutterPlugin, ActivityAware, MethodCallHandler {
                 sourceFileUri = call.argument("sourceFileUri"),
                 sourceFilePath = call.argument("sourceFilePath")
             )
+            "cancelFilesSaving" -> pickOrSave!!.cancelFilesSaving()
             else -> result.notImplemented()
         }
     }